refactor: move get server side props to a shared folder
