+++ src/application/pages/dashboard/utils/get-server-side-props/get-server-side-props.spec.ts
@@ -1,0 +1,7 @@
+import { functionImportTest } from '@utils';
+
+import { getServerSideProps } from './get-server-side-props.function';
+
+describe('getServerSideProps function', () => {
+  functionImportTest(getServerSideProps);
+});
