+++ src/application/pages/dashboard/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './get-server-side-props/get-server-side-props.function';
